Add CWL from CaPTk functionality
