+++ src/cbica_toolkit/cbicaCmdParser.h
@@ -593,7 +593,7 @@ namespace cbica
     {
       argc1ignore = true;
     }
-
+    
   private:
     //! Executable name
     std::string m_exeName;
