+++ src/cbica_toolkit/cbicaUtilities.h
@@ -186,6 +186,8 @@ namespace cbica
   */
   bool exists(const std::string &path);
 
+  std::vector<std::string> getCWLFilesInApplicationDir();
+
   /**
   \brief Create a temporary directory
